Listen History now added
Listen history has been implemented. You can now click Episode History in the navbar and it displays the recent podcasts listened to. Also added fallback images for podcasts and episodes that don't have pictures in the form of the pypods logo. In a random color no less. Lastly, I added the final two tables needed in the database for the upcoming download and queue functionality.