Logging in works - images about 95%
The log in page now works and only accepts logins from users added to the database. However, podcast display currently displays podcasts for all users - instead of just the selected user. This is next on the list and I have a few ideas to fix it.

Images are about 95% of the way to working via web browser and app. In the app images work perfectly. In a browser, only the images hosted locally don't work. This is going on the backburner as it's a very minor issue.