Updates to audio
Audio is mostly finished with flet audio. Should now work in browsers as well. I've introduced true audio caching to allow for quick seeking. In addition made all the audio funtionality work. There's a bug with route changes but I can't fix that. It's a true bug with flet and a issue has been created.