Continually checking off bugs
Really close to a full release. Just checking off the list of bugs. Lots of checks in this commit. It now checks to see if a podcast is already loading when the play button is clicked. Which will prevent errors. In addition, audio volume controls have been implemented. This is almost fully there. Currently the slider is the wrong size. Themes have been finished. Additional verifications have been added as well. There's a popup to tell you a podcast has been removed and it will notify if you try and add the same podcast twice.