Button dynamic updates and pod_view class
2 additions, some of the buttons in settings now update dynamically as you click them. Prior you had to leave the page and come back

I've created a pod_view class. This orients everything oop much better. History is the only page that works with this right now but when you open the page it runs a class and creates an instance for that page that will be used to define the page. This will clean tons of re-used code