Implement settings for OIDC claims & roles
This introduces new settings to tune the OIDC user creation & the
permissions of the users.

---

The first three are to tune the claims (fields in the OIDC user info)
that are used to for the information of the Pinepods user.

`name_claim` can be used to change the claim used for the user's full
name. If not provided this defaults to `name` which was the claim that
was previously used for this. This will now be updated on login,
allowing existing users to be updated if their name changes in the OIDC
provider or if the claim used for this is changed.

`email_claim` can be used to change the claim used for the user's email
address. If not provided this defaults to `email` which was the claim
that was previously always used for this. Note that this is the field
that is used to find an existing account when logging in, so changes to
this claim may cause users to get a new user account in Pinepods.

`username_claim` can be used to change the claim used for the user's
username. If not provided it will use the `preferred_username` claim if
present and fall back to the old behaviour otherwise (using the portion
of the email before the `@` symbol, possibly with a suffix if already in
use). If provided the user's username will be updated on login if
possible, just like is the case for the full name.

---

The next three are to determine the level of access the user should
have. Previoulsy every OIDC user was a regular user in Pinepods, this
allows changing that.

`roles_claim` can be used to change the claim used for the user's roles.
If not provided this defaults to `roles`.

`user_role` is the name of the role that must be present in the list of
roles in order for the user to be allowed to use the application. If the
user doesn't have either this role or the admin role (if set) they will
not be permitted to use Pinepods at all. If this is not set this check
is skipped and all OIDC users will be able to use Pinepods, as was the
case in the past.

`admin_role` is the name of the role that must be present in the list of
roles in order for the user to be granted admin privileges. If the user
has this role during login they will be granted admin rights, and if
they do not their admin rights will be revoked. If this is not set all
users created by logging into OIDC will start out as regular users (as
was the case in the past) and will only have admin access if this is
explictly given to them by another admin in the settings.